backend/requirements.txt
google-genai
fastapi
uvicorn
python-socketio
opencv-python
mediapipe
pillow
mss
playwright
python-dotenv
pyaudio
